+++ src/panels/lovelace/cards/tile/badges/tile-badge-humidifier.ts
@@ -1,31 +1,10 @@
-import {
-  mdiArrowDownBold,
-  mdiArrowUpBold,
-  mdiClockOutline,
-  mdiPower,
-} from "@mdi/js";
 import { stateColorCss } from "../../../../../common/entity/state_color";
 import {
-  HumidifierAction,
+  HUMIDIFIER_ACTION_ICONS,
+  HUMIDIFIER_ACTION_MODE,
   HumidifierEntity,
-  HumidifierState,
 } from "../../../../../data/humidifier";
 import { ComputeBadgeFunction } from "./tile-badge";
-
-export const HUMIDIFIER_ACTION_ICONS: Record<HumidifierAction, string> = {
-  drying: mdiArrowDownBold,
-  humidifying: mdiArrowUpBold,
-  idle: mdiClockOutline,
-  off: mdiPower,
-};
-
-export const HUMIDIFIER_ACTION_MODE: Record<HumidifierAction, HumidifierState> =
-  {
-    drying: "on",
-    humidifying: "on",
-    idle: "off",
-    off: "off",
-  };
 
 export const computeHumidifierBadge: ComputeBadgeFunction = (stateObj) => {
   const hvacAction = (stateObj as HumidifierEntity).attributes.action;
